Refactor Inner Cluster requests.
Create UnitTest for ExternalRequest Handler.
Add extra fields to verify the status of a request.
Fix ClusterMembership unit test, to get rid of other sessions in Zookeeper during testing.

+++ src/main/java/com/grayscaleconsulting/bitacora/data/DataManagerExternal.java
@@ -27,7 +27,7 @@ public interface DataManagerExternal {
      *
      * @param percentageNodesNeededToProceed
      */
-    void setQuorum(long percentageNodesNeededToProceed);
+    void setQuorum(double percentageNodesNeededToProceed);
 
     /**
      * Attempts to cancel a task that is already in progress if another request with the same key
